A bunch of those first-glance type updates
Some new eyes on the code, and finding some new things to fix up. A lot of
documentation type stuff with some code commenting. Make the message and
echo stuff a bit more consistant. Move the UID check later so that we can
see help as a normal user. Other small fixups.

Signed-off-by: Dan McGee <dan@archlinux.org>
Signed-off-by: Simo Leone <simo@archlinux.org>