Remove localtime from overlay rc.conf
We preset HARDWARECLOCK with localtime if the user doesn't set this via
cmdline. But this preset misaligned the BIOS clock if on UTC and the
user select this correctly during aif->date/time-setting.
This is in relation to FS#15263
Again: There is no need in LiveCD environment to modify BIOS clock
during shutdown.
Having this set per default to localtime rc.shutdown sources rc.conf and
so localtime is badly used. And worked against the meaning of above
flyspray report.
(Thanks fs4000, for point me on that)

Signed-off-by: Gerhard Brauer <gerbra@archlinux.de>
Signed-off-by: Aaron Griffin <aaronmgriffin@gmail.com>