[archiso] Introduce cowfile_size= bootparam
If not set, nothing is changed in current behaviour.

Since dm-snapshot allows to use as "COW" a file with any size,
without caring about the the size of "origin", we can avoid creating a
"COW" file of the same size as the "origin". This is really useful,
when using as cow_device= a filesystem that is VFAT where sparse files
are not supported, so if root-image.fs is 1000M, passing cowfile_size=25%
will create a root-image.cow of 250M instead of 1000M.

Signed-off-by: Gerardo Exequiel Pozzi <vmlinuz386@yahoo.com.ar>