Unsetting the root password
Due to changes in the filesystem package [1] the current /etc/passwd
does not allow login without an empty root password anymore.
This becomes apparent, when switching to another TTY than the one
autologin is activated for.

By running passwd -d root we unset the password.
In the future this should go to a custom /etc/passwd file.

[1] https://git.archlinux.org/svntogit/packages.git/commit/trunk?h=packages/filesystem&id=0320c909f3867d47576083e853543bab1705185b